+++ colorpicker/src/main/kotlin/ca/allanwang/kau/colorpicker/CircleView.kt
@@ -45,9 +45,10 @@ class CircleView @JvmOverloads constructor(context: Context, attrs: AttributeSet
     private val outerPaint: Paint = Paint().apply { isAntiAlias = true }
     private val whitePaint: Paint = Paint().apply { isAntiAlias = true; color = Color.WHITE }
     private val innerPaint: Paint = Paint().apply { isAntiAlias = true }
+    val colorSelected: Boolean
+        get() = selected
     private var selected: Boolean = false
     var withBorder: Boolean = false
-        get() = field
         set(value) {
             if (field != value) {
                 field = value
@@ -65,13 +66,13 @@ class CircleView @JvmOverloads constructor(context: Context, attrs: AttributeSet
         outerPaint.color = shiftColorDown(color)
 
         val selector = createSelector(color)
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+        foreground = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
             val states = arrayOf(intArrayOf(android.R.attr.state_pressed))
             val colors = intArrayOf(shiftColorUp(color))
             val rippleColors = ColorStateList(states, colors)
-            foreground = RippleDrawable(rippleColors, selector, null)
+            RippleDrawable(rippleColors, selector, null)
         } else {
-            foreground = selector
+            selector
         }
     }
 
@@ -85,20 +86,17 @@ class CircleView @JvmOverloads constructor(context: Context, attrs: AttributeSet
         setBackgroundColorRes(color)
     }
 
-
-    @Deprecated("")
+    @Deprecated("Cannot use setBackground() on CircleView", level = DeprecationLevel.ERROR)
     override fun setBackground(background: Drawable) {
         throw IllegalStateException("Cannot use setBackground() on CircleView.")
     }
 
-
-    @Deprecated("")
+    @Deprecated("Cannot use setBackgroundDrawable() on CircleView", level = DeprecationLevel.ERROR)
     override fun setBackgroundDrawable(background: Drawable) {
         throw IllegalStateException("Cannot use setBackgroundDrawable() on CircleView.")
     }
 
-
-    @Deprecated("")
+    @Deprecated("Cannot use setActivated() on CircleView", level = DeprecationLevel.ERROR)
     override fun setActivated(activated: Boolean) {
         throw IllegalStateException("Cannot use setActivated() on CircleView.")
     }
@@ -111,25 +109,16 @@ class CircleView @JvmOverloads constructor(context: Context, attrs: AttributeSet
 
     fun animateSelected(selected: Boolean) {
         if (this.selected == selected) return
-        this.selected = true // We need to draw the other bands
+        this.selected = selected // We need to draw the other bands
         val range = if (selected) Pair(-borderWidthSmall, borderWidthLarge) else Pair(borderWidthLarge, -borderWidthSmall)
-        val animator = ValueAnimator.ofFloat(range.first, range.second)
-        with(animator) {
+        ValueAnimator.ofFloat(range.first, range.second).apply {
             reverse()
             duration = 150L
             addUpdateListener { animation ->
                 whiteOuterBound = animation.animatedValue as Float
                 invalidate()
-            }
-            addListener(object : AnimatorListenerAdapter() {
-                override fun onAnimationEnd(animation: Animator) {
-                    this@CircleView.selected = selected
-                }
 
-                override fun onAnimationCancel(animation: Animator) {
-                    this@CircleView.selected = selected
-                }
-            })
+            }
             start()
         }
     }
@@ -194,10 +183,10 @@ class CircleView @JvmOverloads constructor(context: Context, attrs: AttributeSet
         cheatSheet.show()
     }
 
-    companion object {
+    private companion object {
 
         @ColorInt
-        private fun translucentColor(color: Int): Int {
+        fun translucentColor(color: Int): Int {
             val factor = 0.7f
             val alpha = Math.round(Color.alpha(color) * factor)
             val red = Color.red(color)