+++ core/src/main/kotlin/ca/allanwang/kau/utils/NetworkUtils.kt
@@ -1,5 +1,6 @@
 package ca.allanwang.kau.utils
 
+import android.annotation.SuppressLint
 import android.content.Context
 import android.net.ConnectivityManager
 
@@ -7,6 +8,7 @@ import android.net.ConnectivityManager
  * Created by Allan Wang on 2017-07-07.
  */
 inline val Context.isNetworkAvailable: Boolean
+    @SuppressLint("MissingPermission")
     get() {
         val connectivityManager = getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
         val activeNetworkInfo = connectivityManager.activeNetworkInfo
@@ -14,6 +16,7 @@ inline val Context.isNetworkAvailable: Boolean
     }
 
 inline val Context.isWifiConnected: Boolean
+    @SuppressLint("MissingPermission")
     get() {
         val connectivityManager = getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
         val activeNetworkInfo = connectivityManager.activeNetworkInfo
@@ -21,6 +24,7 @@ inline val Context.isWifiConnected: Boolean
     }
 
 inline val Context.isMobileDataConnected: Boolean
+    @SuppressLint("MissingPermission")
     get() {
         val connectivityManager = getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
         val activeNetworkInfo = connectivityManager.activeNetworkInfo